agente_app/requirements.txt
flask
google-generativeai
google-adk
psycopg2-binary
